Apply authentication on outgoing requests
Requests had to be created using the `request_builder` helper, which
applied authentication ahead of time. This is not suitable for
authentication schemes which need the actual request to be finalised
first (e.g.: Digest auth).

This also resulted in a quirky API, where requests couldn't be built
normally. The API also didn't allow submitting ad-hoc HTTP requests to
the server.

This change moves the logic applying authentication to the point where
requests are about to be dispatched, which means they're already
complete (e.g.: for things like Digest auth) and can be build normally
via `Request::builder()`. This also now allows sending out arbitrary
requests.

Finally, this changes the authentication API a bit to better fit this
scheme, and introduces a new error type to be used in any call-path that
sends out WebDav requests.