Refactor synchronisation implementation
The main goal here was to improve the separation between declaration,
planning and executing. These are not in three separate files, and they
each only import the module for the previous step, but not the next.

This also introduces a support for mapping collections across storages,
with a strong focus on providing a clean API to defined these mappings.
This API can be used by various frontends, including vdirsyncer itself.

The internal implementation could still use some improvements,
especially on the performance side. There are also some minor issues
with invalid input. Mainly: specifying mappings for the same collection
more than once will result in a failed synchronisation, rather than
reject the input entirely.