libdav: make bootstrap part of the builder sequence
Previously one had to create a client _and then_ execute the discovery
(a.k.a.: bootstrap) process, which implied mutating the client itself.
Mutating the client is awkward, especially given that we're using a
builder in the first place.

This makes both clients immutable, including changing the visibility to
fields that can now be private.