Remove `mut` from write operations on Storage
My original intent was that any write operation on a Storage requires a
`mut` reference to it. This could prevent having two copies of a Storage
that concurrently apply modifications.

Experience proves that this approach wasn't a good idea and is an abuse
of the variable mutability.

To begin with, `mut` is intended to indicate that a function mutates a
variable (e.g.: the in-memory representation itself). These methods
don't mutate the storage itself; they execute mutations on the remote
server that it represents (or on the filesystem). This is much like a
`POST` request doesn't mutate an HTTP client, it executes an operation
which causes a mutation on the server.

The different between both might sound subtle, but it's not.

Storage implementations themselves ought to be atomic too, so actually
preventing concurrent operations which mutate the storage shouldn't be
needed. For example, writing is atomic, so if two concurrent operations
are executed, one ought to fail. There should never be undefined
behaviour due to concurrent operations (if there are, that is a bug in
the implementation).

A second problematic point with the existing approach is that it won't
allow executing two concurrent operations at all. When uploading 800
calendar events, they each need to be uploaded sequentially. This is
clearly terrible.