Use Arc to handle multiple references to Storage
Trying to juggle multiple references turns into a pain with container
structures that want to own both the storage and a pair (which is what
vdirsyncer needs to do).

Using an Arc is arguably cleaner than shared references to a Box, and
will likely be a lot more ergonomic to library consumers.