Prefer `impl Trait` for parameter types
This is done mostly so a few functions take IntoIter<S> instead of &[S],
which avoids the need to actually allocate an array. The rest of the
usages are updated for consistency.

Also improves clarity/readability of several APIs.

See: https://github.com/rust-lang/rfcs/blob/master/text/1951-expand-impl-trait.md#motivation-for-expanding-to-argument-position