Drop the Definition trait
The original intent of this trait was to have a deserialiseable
definition of each storage type, which could be read from a
configuration file and converted into a storage.

The current implementation failed in this respect, given that had types
that were not trivially deserialisable. Storage definitions can include
rules for TLS configuration, and deserialising those is beyond the scope
of this crate, so adapting definitions to actually fit their original
goal is not really a solid design.

Drop the Definition trait entirely, and expose direct constructors for
each Storage type. This also results in a simpler API.

Fixes: https://todo.sr.ht/~whynothugo/vdirsyncer-rs/42