Refactor sync planing
This drops a lot of unnecessary logic from the old state-based
implementation.

The sync::state module is gone entirely, and only CollectionState state
remains, which is a private type of the sync::plan module now.

As a result, the logic in planning is a lot more collection-oriented in
its flow.