Rewrite constructor/bootstrap interface for libdav
The builder interface being used to create new clients is too
complicated. Usability was more complicated than it needed to be, and
the internal implementation was a huge mess. Most of the sub-portions of
discovery were not easily usable outside the main bootstrapping flow

Replace it entirely with a simpler `fn new` function for creating
clients, and `fn new_via_bootstrap` for creating a new client using
automatic bootstrap.

The basic flow to create a new CalDavClient now is:

- Create an `Auth` instance.
- Create a (https?) connector.
- Create a `WebDav` client.
- Create a `CalDavClient`.

Clients no longer have a `principal` field; a principal may or may not
exist, but is data external to the client that can be fetched explicitly
when needed.

A large portion of these changes consist of rewriting call-sites using
the refactored interface.