A home set may return multiple hrefs
I mis-interpreted the spec previously; the home set is actually a set of
collections, not a collection that contains a set of calendars.

This commit implements support for a home set with multiple collections
in libdav and davcli, but vstorage and live_tests still need to be
adapted.

Fixes: https://todo.sr.ht/~whynothugo/vdirsyncer-rs/60
See: https://todo.sr.ht/~whynothugo/vdirsyncer-rs/61
See: https://todo.sr.ht/~whynothugo/vdirsyncer-rs/62