Switch to scfg configuration format
I don’t really like toml as a format, and the main reason for using it
so far was that it results in minimal change from v1.

While change for the sake of change is problematic, change to provide a
simpler configuration syntax is an improvement. I don’t want to release
v2 and regret having committed to a format that I dislike. I also don't
want to change the configuration format in incompatible ways once a
stable release has been made.

Given than this release requires updating the configuration file anyway,
this is the ideal opportunity to adopt what I consider the ideal choice.

Fear of change is almost as bad as change for the sake of change.

---

Aside from changing the format itself, this cleans up a lot of the code
related to parsing configuration files. Configuration issues in disabled
pairs and storages won't block vdirsyncer from proceeding.