Ignore absent resources when parsing results
Results which return 404 don't have an Etag, the result fails to parse,
and a bogus error ("invalid data").

Ignore items in the response which are not found or some other
non-success code. This is a better match for the intended API, which
should just list items which are found.

If a single item is included in the response, return a DoesNotExist
error. This is somewhat of a band-aid, and should be improved upon in
future.