Finalise support for Dav over Unix sockets
Allow usage of Dav servers directly over Unix domain sockets. The
enables a variety of scenarios, such as using a local server without
exposing it to the network, or using proxies which handle custom
authentication, transports or TLS configurations.

When a storage URL starts with unix://, support is automatically
enabled. An example usage is using the configuration directive:

    url unix:///run/dav/socket

Documentation is pending, and falls within the scope of the rest of the
general documentation.

See: https://whynothugo.nl/journal/2024/08/19/unix-domain-socket-support-for-vdirsyncer/
Closes: https://todo.sr.ht/~whynothugo/vdirsyncer-rs/36