Use pre-fetched data when creating items
The planning process fetches all new items and produces a plan. The
execution process then fetches items that needed to be copied, one at a
time.

Fetching items individually is slow, especially over the network where
each item requires an individual HTTP request. Additionally, items are
already fetched in bulk during the planning process.

Include the pre-fetched data in a plan, and re-use that when executing
it.

This dramatically improves performance when creating new items fetched
from a remote source.

Fixes: https://todo.sr.ht/~whynothugo/vdirsyncer-rs/101