Fix crash if a previous run left an sqlite journal file
If a previous execution left behind an sqlite journal file, the database
cannot be opened in read-only mode, because sqlite needs to apply the
incomplete journal (or roll it back) before it can open the database in
a read-only way that is thread-safe.

Always open the database in read-write mode, but retain a function to
open it without creating it.

Fixes: https://todo.sr.ht/~whynothugo/pimsync/164