Re-initialise inotify in case of failure
After a directory/collecitons is deleted, inotify will immediately
return an error when polled. This continuous stream of errors result in
a continuous stream of events which in turn trigger the sync process
repeatedly and indefinitely.

Handle inotify having failed. Re-initialise it if possible, or fall back
to the timer if reinitialisation is not possible.