Implement subcommand to repair items
Storages may have items invalid in a variety of ways. A common scenario
is for items to lack a UID. Implement a basic routine to fix such items
by adding a UID to them.

Concurrency had to be partially stripped out, due to risk of deadlock
and risk of overwhelming storages with concurrent requests.