create_item: allow specifying names
We want to preserve filenames when syncing between compatible storages.

The provided name is used if valid for the given storage; otherwise a
fallback is used.

Use a new type so that future optional arguments can be included inside
of it, and to simplify usage via CreateItemOptions::default().

See: https://todo.sr.ht/~whynothugo/pimsync/88